Change the method which called by H2 Dao implements.
