+++ apm-collector/apm-collector-boot/src/main/resources/application.yml
@@ -27,16 +27,10 @@ collector_inside:
   grpc:
     host: localhost
     port: 11800
-storage:
-  elasticsearch:
-    cluster_name: CollectorDBCluster
-    cluster_transport_sniffer: true
-    cluster_nodes: localhost:9300
-    index_shards_number: 2
-    index_replicas_number: 0
-# uncomment to enable h2 storage
 #storage:
-#    h2:
-#      url: jdbc:h2:~/collector
-#      user_name: sa
-#      password: sa
+#  elasticsearch:
+#    cluster_name: CollectorDBCluster
+#    cluster_transport_sniffer: true
+#    cluster_nodes: localhost:9300
+#    index_shards_number: 2
+#    index_replicas_number: 0
