+++ apm-collector/apm-collector-ui/src/main/java/org/skywalking/apm/collector/ui/dao/ServiceEntryH2DAO.java
@@ -18,11 +18,13 @@
 
 package org.skywalking.apm.collector.ui.dao;
 
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
-
+import org.skywalking.apm.collector.cache.ApplicationCache;
 import org.skywalking.apm.collector.client.h2.H2Client;
 import org.skywalking.apm.collector.client.h2.H2ClientException;
 import org.skywalking.apm.collector.core.util.ColumnNameUtils;
@@ -30,12 +32,8 @@ import org.skywalking.apm.collector.core.util.StringUtils;
 import org.skywalking.apm.collector.storage.define.service.ServiceEntryTable;
 import org.skywalking.apm.collector.storage.h2.SqlBuilder;
 import org.skywalking.apm.collector.storage.h2.dao.H2DAO;
-import org.skywalking.apm.collector.ui.cache.ApplicationCache;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import com.google.gson.JsonArray;
-import com.google.gson.JsonObject;
 
 /**
  * @author pengys5
@@ -43,6 +41,7 @@ import com.google.gson.JsonObject;
 public class ServiceEntryH2DAO extends H2DAO implements IServiceEntryDAO {
     private final Logger logger = LoggerFactory.getLogger(SegmentH2DAO.class);
     private static final String GET_SERVICE_ENTRY_SQL = "select * from {0} where {1} >= ? and {2} <= ?";
+
     @Override public JsonObject load(int applicationId, String entryServiceName, long startTime, long endTime, int from,
         int size) {
         H2Client client = getClient();
@@ -77,7 +76,7 @@ public class ServiceEntryH2DAO extends H2DAO implements IServiceEntryDAO {
             while (rs.next()) {
                 int appId = rs.getInt(ServiceEntryTable.COLUMN_APPLICATION_ID);
                 int entryServiceId = rs.getInt(ServiceEntryTable.COLUMN_ENTRY_SERVICE_ID);
-                String applicationCode = ApplicationCache.getForUI(applicationId);
+                String applicationCode = ApplicationCache.get(applicationId);
                 String entryServiceName1 = rs.getString(ServiceEntryTable.COLUMN_ENTRY_SERVICE_NAME);
 
                 JsonObject row = new JsonObject();
